pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "pydocs-mcp"
version = "0.1.0"
description = "Offline Python and ML documentation MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "selectolax>=0.3",
  "markdownify>=0.13",
  "PyYAML>=6.0",
  "pydantic>=2.7",
  "typer>=0.12",
  "mcp>=1.0,<2",
]

[project.scripts]
pydocs-mcp = "pydocs_mcp.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pydocs_mcp"]

[tool.hatch.build]
include = ["src/pydocs_mcp/data/*.yaml"]
